delete SessionService
